enhancement: deduplicate pkg.ts validation logic between cloud-cli and strapi
