+++ packages/core/utils/src/index.ts
@@ -39,3 +39,4 @@ export * from './route-serialization';
 export * from './primitives';
 export * from './content-api-router';
 export * from './security';
+export * from './pkg-utils';
